pyproject.toml
[project]
name = "realtime-translator"
version = "0.4.3"
description = "Real-time audio translation using Whisper + LLM"
requires-python = ">=3.11"
dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.24",
    "faster-whisper>=1.1.0",
    "PyQt6>=6.6.0",
    "pyyaml>=6.0",
    "openai>=1.30.0",
    "huggingface-hub>=0.20.0",
    "pyaudiowpatch>=0.2.12",
    "pyannote.audio>=4.0.4",
]

[project.scripts]
realtime-translator = "main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pyinstaller>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "ui", "utils"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
